Add docblocks to all classes and interfaces
This is need to work around a PHP bug resulting in a (random?)
docblock being returned when calling getDocComment() on a class
reflection of a class that has no docblock.

See PHP bug 55156. Although reported to be fixed as of 5.3.7 the
test for the bug fails in 5.3.8 (for 5.4.0 the test passes).

The bug then causes errors to be reported by Doctrine's annotation
reader, because it "sees" annotations on classes that don't have
them, and thus lack needed use statements. Bummer.