[WIP][BUGFIX] Improve memory consumption for thumbnails
When creating a thumbnail, no longer a copy of the initial,
probably very large image is created, but the resizing
happens directly on the initial image.

For convenience, two new `BoxInterface` methods have been
introduced, `inset()` and `outbound()`, which will provide
an according resizing of a `Box`. The `Image`'s `thumbnail`
method has been refactored to make use of these methods. Credits
go to @noginn for the idea and the code snippet.

Fixes #149