Reverse deprecated version logic.
Imagick does not have a version contained inside it, unless it was packaged by PECL. For versions compiled from source, the version is not a semver number, and so can't be compared sanely. This change makes the code assume that Imagick is being compiled from source, then it is from post 3.1.0b1, and not some code that was released on in February 2011.